refactor: Manual content lives on the source (config_json.graph)
The user pushed back on "attach is a prerequisite to edit Manual":
"なぜ attach 前提なの？？". Right — Manual fundamentally has no
upstream, so its "content" IS the source itself, not something
produced inside a topology context.

Move Manual graph from `topology_observations` (keyed by topology, source)
to `data_sources.config_json` under a `graph` key (source-level).
The same Manual graph is now visible to every topology it 's attached
to. Orphan Manuals (created but not attached) now make sense — they
have content waiting to be wired in.

[Storage]
- Migration 011: for each Manual data_source, take its most-recent
  observation and stash that graph into config_json.graph. Then
  delete topology_observations rows for Manual sources.

[Server]
- `TopologyService.attachManualSource`: stops recording an empty
  observation; seeds config_json with `{"graph": {empty graph}}`
  instead. Manual stays per-topology at the *attach* level (one Manual
  per topology, by current cardinality choice) but its content is
  source-level.
- `TopologyService.parseTopology`: reads Manual graph from
  `data_sources.config_json.graph` via new `readManualGraph(sourceId)`
  helper. Non-Manual sources still flow through observations
  unchanged. NetBox / SNMP behaviour is untouched.

[UI — editor moves back to /datasources/<manualId>]
- The dedicated `/topologies/[id]/sources/[sourceId]/edit` route is
  deleted. The (topology, source) editor pattern made sense for
  per-(topology, source) content, which Manual no longer has.
- `/datasources/[id]/+page.svelte` regains a YAML/JSON graph editor
  for `type === "manual"`. Save serialises the config as
  `{"graph": <NetworkGraph>, ...}` and PUTs to the standard
  datasource update endpoint — no new server endpoint needed.
- Attached-topology list still renders below the editor as a
  "Used by: <topology>…" hint pointing at /topologies/<id>.
- All Edit-Manual links updated:
  - /topologies/+page.svelte create modal post-redirect
  - /topologies list row Edit button
  - /topologies/[id]/settings Actions card
  - TopologySettings drawer Actions
  All now point to `/datasources/<manualSourceId>`.

[UI — settings edgeStyle / splineMode]
These live on `NetworkGraph.settings` which now sits inside
config_json.graph. Switched `parseGraphSettings` and `updateEdgeStyle`
in both the settings page and the drawer to read / write through
`api.dataSources.get` / `dataSources.update`. The latestSnapshot /
recordObservation calls these used are now dead-letter for Manual
(observations don 't drive Manual rendering anymore).

[Note] The `api.topologies.sources.latestSnapshot` /
`recordObservation` / `attachManual` helpers are still in `lib/api.ts`
because they remain useful for non-Manual sources (NetBox / SNMP
observations stay observation-shaped). Just nothing in the Manual
path calls them anymore.

Validation: bun run typecheck 34/34, svelte-check 0/0.
