+++ apps/server/api/src/db/schema.ts
@@ -16,6 +16,7 @@ import migration006 from './migrations/006_share_tokens.sql'
 import migration007 from './migrations/007_grafana_alerts.sql'
 import migration008 from './migrations/008_topology_observations.sql'
 import migration010 from './migrations/010_manual_as_source.sql'
+import migration011 from './migrations/011_manual_graph_to_config.sql'
 
 /** Ordered list of all migrations */
 const MIGRATIONS: { name: string; sql: string }[] = [
@@ -28,6 +29,7 @@ const MIGRATIONS: { name: string; sql: string }[] = [
   { name: '007_grafana_alerts.sql', sql: migration007 },
   { name: '008_topology_observations.sql', sql: migration008 },
   { name: '010_manual_as_source.sql', sql: migration010 },
+  { name: '011_manual_graph_to_config.sql', sql: migration011 },
 ]
 
 interface MigrationRecord {
