+++ apps/server/web/src/lib/components/TopologySettings.svelte
@@ -5,6 +5,7 @@
   import { api } from '$lib/api'
   import { Button } from '$lib/components/ui/button'
   import {
+    dataSources,
     displaySettings,
     liveUpdatesEnabled,
     metricsConnected,
@@ -33,15 +34,16 @@
   let splineMode = $state('sloppy')
 
   // edgeStyle / splineMode live on `NetworkGraph.settings` inside the
-  // Manual source 's snapshot. We route reads/writes through the
-  // per-source observation endpoint — same pattern as the dedicated
-  // Manual editor page.
+  // Manual source 's graph (config_json.graph). Read/write via the data
+  // source config — same path as the Manual editor page.
   async function parseGraphSettings() {
     if (!topology.manualSourceId) return
     try {
-      const snap = await api.topologies.sources.latestSnapshot(topology.id, topology.manualSourceId)
-      const settings = (snap.graph as { settings?: { edgeStyle?: string; splineMode?: string } })
-        ?.settings
+      const ds = await api.dataSources.get(topology.manualSourceId)
+      const config = JSON.parse(ds.configJson || '{}') as {
+        graph?: { settings?: { edgeStyle?: string; splineMode?: string } }
+      }
+      const settings = config.graph?.settings
       edgeStyle = settings?.edgeStyle || 'orthogonal'
       splineMode = settings?.splineMode || 'sloppy'
     } catch {
@@ -53,8 +55,12 @@
     if (!topology.manualSourceId) return
     savingEdgeStyle = true
     try {
-      const snap = await api.topologies.sources.latestSnapshot(topology.id, topology.manualSourceId)
-      const graph = (snap.graph ?? { version: '1', nodes: [], links: [] }) as unknown as {
+      const ds = await api.dataSources.get(topology.manualSourceId)
+      const config = JSON.parse(ds.configJson || '{}') as {
+        graph?: { settings?: Record<string, unknown>; [k: string]: unknown }
+        [k: string]: unknown
+      }
+      const graph = (config.graph ?? { version: '1', nodes: [], links: [] }) as {
         settings?: Record<string, unknown>
         [k: string]: unknown
       }
@@ -65,11 +71,10 @@
       } else {
         delete graph.settings['splineMode']
       }
-      await api.topologies.sources.recordObservation(
-        topology.id,
-        topology.manualSourceId,
-        graph as unknown as import('@shumoku/core').NetworkGraph,
-      )
+      config.graph = graph
+      await dataSources.update(topology.manualSourceId, {
+        configJson: JSON.stringify(config),
+      })
       if (onUpdated) onUpdated(topology)
     } catch (e) {
       console.error('Failed to update edge style:', e)
@@ -403,7 +408,7 @@
     <div class="space-y-2">
       {#if topology.manualSourceId}
         <a
-          href="/topologies/{topology.id}/sources/{topology.manualSourceId}/edit"
+          href="/datasources/{topology.manualSourceId}"
           class="btn btn-secondary w-full justify-center"
         >
           <PencilSimpleIcon size={16} class="mr-2" />
